FIX : int > Long 타입변경

+++ src/main/java/com/nawabali/nawabali/dto/UserDto.java
@@ -31,8 +31,8 @@ public class UserDto {
         String city;
         String district;
         UserRankEnum rank;
-        int totalLikesCount;
-        int totalLocalLikesCount;
+        Long totalLikesCount;
+        Long totalLocalLikesCount;
         String profileImageUrl;
